pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "conversational-voice-agent"
version = "0.1.0"
description = "マイク入力からリアルタイムで音声会話できるAIエージェント"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "anthropic>=0.40.0",
    "faster-whisper>=1.0.0",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
voice-agent = "voice_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_agent"]
